+++ carini/carini/src/main/java/com/car/controller/MypageController.java
@@ -24,7 +24,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Controller
-public class mypageController {
+public class MypageController {
 
 	@Autowired
 	private BookMarkService bookMarkService;
