+++ carini/carini/src/main/java/com/car/impl/BoardServiceImpl.java
@@ -1,5 +1,6 @@
 package com.car.impl;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,12 +9,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.car.dto.Board;
+import com.car.dto.Car;
+import com.car.dto.Member;
 import com.car.persistence.BoardRepository;
 import com.car.service.BoardService;
 
 @Service
 public class BoardServiceImpl implements BoardService{
-	
+
 	@Autowired
 	private BoardRepository boardRepository;
 	
@@ -62,4 +65,21 @@ public class BoardServiceImpl implements BoardService{
 	public void deleteBoard(Board board) {
 		boardRepository.deleteById(board.getBoardId());
 	}
+	
+	
+	@Override
+	public List<Board> boardList(Member member) {
+		
+		List<Board> boardList = boardRepository.findBymemberId(member.getMemberId());
+		
+		if(boardList.isEmpty()) {
+			System.out.println("66666");
+			return null;
+		}else {
+			return boardList;
+		}
+	}
+
+	
+	
 }
