+++ carini/carini/src/main/java/com/car/impl/BookMarkServiceImpl.java
@@ -18,7 +18,7 @@ import com.car.service.BookMarkService;
 import jakarta.transaction.Transactional;
 
 @Service
-public class BookMarkServiceimpl implements BookMarkService{
+public class BookMarkServiceImpl implements BookMarkService{
 
 	@Autowired
 	private BookMarkRepository bookMarkRepository;
